requirements.txt
finnhub-python
yfinance
openai
jinja2
python-dotenv
